Add files via upload
Fixed a bug, in which commercial demand was influenced partly by industrial zoning, and industrial demand was influenced partly by the population count. Now, the commercial demand is influenced partly by the population count, and industrial demand is influenced partly by the commercial zoning. Additionally, a quick fix aiming for the residential, commercial and industrial demand rebalance is present.